requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
joblib==1.4.2
numpy==2.0.1
nvidia-nccl-cu12==2.22.3
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
xgboost==2.1.1
